SpatialHashing

Documentation for SpatialHashing.

SpatialHashing.each_near

SpatialHashing.spatial_hash

SpatialHashing.spatial_hash!

SpatialHashing.update!

SpatialHashing.each_near — Method

SpatialHashing.each_near(fun,x,search_range,spatial_index,visited)

Evaluates fun for every index j in points which is near x. The search is limited to ± spatial_index around the grid cell containing x. The vector visited is an vector of booleans of the size points to keep track which points have been found so far.

It is important that the function fun checks that the proposed point with index j is actually close as different grid cells may have the same hash (hash-collision).

The following is a complete example in 2 dimensions:

npoints = 200
points = [Tuple(rand(2)) for i = 1:npoints]
h = 0.1
max_table = 50
spatial_index = SpatialHashing.spatial_hash(points,h,max_table)
x = (0.2,0.2)
r2max = 0.1^2
search_range = 1
visited = falses(npoints)
SpatialHashing.each_near(x,search_range,spatial_index,visited) do j
   r2 = sum((points[j] .- x).^2)
   if r2 < r2max
      println("point with index ",j," is near (distance is ",sqrt(r2),")")
   end
end

# the same indices as naive search

filter(i -> sum((points[i] .- x).^2) < r2max,1:length(points))

SpatialHashing.spatial_hash! — Method

SpatialHashing.spatial_hash!(points,h,table,num_points)

Initialize the data structure for spatial hashing in-place. table is a vector of integers with hashes, num_points is a vector of integers with the same length as points.

See SpatialHashing.spatial_hash for explanations of the other parameters.

SpatialHashing.spatial_hash — Method

spatial_index = SpatialHashing.spatial_hash(points,h,max_table)

Initialize the data structure spatial_index for spatial hashing using the vector of position points and grid with resolution h. max_table is the length of the hash table (while size can be tuned to optimize the performance).

The index allows to localize all points near a given reference point in O(log(n)) operations where n is the number of points.

The hash function is currently setup for 1- up to 6-dimensional spaces.

The parameter h is used to define a grid of cell. Each particle belongs to one grid cell. For every cell a hash number is computed and the index keeps track of all points with the same hash.

   <──h──>
   ┏━━━━━━┯━━━━━━┯━━━━━━┯━━━━━━┯━━━━━━┯━━━━━━┓
   ┃      │      │      │      │      │      ┃
   ┃      │      │      │      │      │      ┃
   ┠──────┼──────┼──────┼──────┼──────┼──────┨
   ┃  ①   │      │      │      │      │      ┃
   ┃      │      │      │      │      │      ┃
   ┠──────┼──────┼──────┼──────┼──────┼──────┨
   ┃      │      │      │  ④   │      │      ┃
   ┃      │      │      │      │      │      ┃
   ┠──────┼──────┼──────┼──────┼──────┼──────┨
   ┃      │  ②   │      │      │      │      ┃
   ┃      │      │      │      │      │      ┃
   ┠──────┼──────┼──────┼──────┼──────┼──────┨
   ┃      │      │      │      │    ⑤ │      ┃
   ┃      │      │      │      │ ③    │      ┃
   ┗━━━━━━┷━━━━━━┷━━━━━━┷━━━━━━┷━━━━━━┷━━━━━━┛
 (0,0)

SpatialHashing.update! — Method

SpatialHashing.update!(spatial_index,points)

Update the spatial spatial_index using the new positions of the points. The number of points is assumed to be the same as during initialization with SpatialHashing.spatial_hash.